rename example to zeticai
