+++ whisper/ZeticMLangeWhisper-Android/app/src/test/java/com/zeticai/whisper/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.whisper_demo_app
+package com.zeticai.whisper
 
 import org.junit.Test
